[bugged]: Introduce canDeleteAnyContainer flag and use LazyColumn
This commit introduces a new global flag `canDeleteAnyContainer` in `HomeViewModel` to determine if any container can be deleted. This logic, previously handled by checking the list size, is now managed by this flag and used in `PojedynczyKontenerWalutyUI` to control swipe-to-delete functionality.

The `GlownyEkran` now uses `LazyColumn` for displaying currency containers, replacing the previous `Column` with `verticalScroll`. Each container is rendered by the new `PojedynczyKontenerWalutyUI` composable, which also handles the swipe-to-delete animation and logic based on the `canBeSwipedToDelete` flag.

Additionally, library versions for `lifecycleRuntimeCompose`, `composeBom`, and `firebaseStorageKtx` have been updated.