Feat: Add AD_ID permission for Firebase Analytics
This commit adds the `com.google.android.gms.permission.AD_ID` permission to the `AndroidManifest.xml`. This permission is required by Firebase Analytics to collect advertising identifiers, which can be used for more accurate user measurement and attribution.

The deployment target was also updated from a physical device to a local emulator.